Move App\Extensions\Middleware\ContentGuard to namespace App\Extensions\Http
